fix: 계획 저장, 삭제 서비스 수정
<계획 저장>
1. 여행명, 여행기간, 이미지 저장
2. 출발위치(MapStatus.START), 도착위치(MapStatus.END)는 MAP 테이블에 따로 저장
3. 이미지는 도착위치 이미지로 저장

<계획 삭제>
1. @AuthenticationPrincipal 사용해서 계획 삭제

+++ src/main/java/be/triplan/service/PlanService.java
@@ -1,7 +1,6 @@
 package be.triplan.service;
 
 import be.triplan.dto.plan.PlanDto;
-import be.triplan.dto.plan.PlanUpdateRequestDto;
 import be.triplan.entity.Member;
 import be.triplan.entity.Plan;
 import be.triplan.entity.PlanJoin;
@@ -66,7 +65,13 @@ public class PlanService {
 
     //계획 삭제
     @Transactional
-    public void delete(Long id) {
-        planRepository.deleteById(id);
+    public void delete(Member member, Long id) {
+        List<PlanJoin> planJoins = planJoinRepository.findByMember_Id(member.getId());
+        PlanJoin planJoin = planJoins.get(0);
+
+        Long planJoinPlanId = planJoin.getPlan().getId();
+        if (planJoinPlanId == id) {
+            planRepository.deleteById(id);
+        }
     }
 }
